Reorganize fixtures
