+++ projects/07/vm_translator/tests/test_parser.py
@@ -1,48 +1,8 @@
-from io import StringIO
-
 import pytest
 
 from parser import (
-    ArithmeticCommandClass,
-    Command,
-    Parser,
-    StackCommandClass
+    Parser
 )
-
-
-@pytest.fixture
-def file_with_comments():
-    initial_value = '// This is a test\n// And nothing is here'
-    return StringIO(initial_value=initial_value)
-
-
-@pytest.fixture
-def file_with_empty_lines():
-    initial_value = '\n    \n\t\n'
-    return StringIO(initial_value=initial_value)
-
-
-@pytest.fixture(params=ArithmeticCommandClass)
-def file_with_arithmetic_command(request):
-    dummy_file_obj = StringIO(
-        initial_value=f"{request.param.value}"
-    )
-    return dummy_file_obj, request.param
-
-
-@pytest.fixture(params=StackCommandClass)
-def file_with_stack_command(request):
-    dummy_file_obj = StringIO(
-        initial_value=f"{request.param.value} segment 0"
-    )
-    return dummy_file_obj, request.param
-
-
-@pytest.fixture
-def file_with_garbage():
-    return StringIO(
-        initial_value='this is nothing like a command'
-    )
 
 
 def test_discards_comments(file_with_comments):
